bazel: Split packages up
Rather than having one package with all sources, make each Java/Kotlin
package its own bazel package. Then they can be built in parallel if
they don't depend on each other. And if they do depend on each other, we
can avoid rebuilds of packages at the top of the dependency tree, if
they haven't changed.
