+++ src/me/danielschaefer/sensorwrangler/Library.kt
@@ -1,0 +1,5 @@
+package me.danielschaefer.sensorwrangler
+
+// Empty library because kt_jvm_library does not support empty srcs
+// See: https://github.com/bazelbuild/rules_kotlin/issues/232
+class Library
