+++ src/me/danielschaefer/sensorwrangler/annotations/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_kotlin//kotlin:jvm.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "annotations",
+    srcs = glob([ "*.kt" ]),
+    deps = [
+        "//:lib_deps",
+    ],
+    visibility = [
+        "//src/me/danielschaefer/sensorwrangler/javafx:__pkg__",
+        "//src/me/danielschaefer/sensorwrangler/sensors:__pkg__",
+    ],
+)
