+++ src/me/danielschaefer/sensorwrangler/data/Picker.kt
@@ -1,0 +1,5 @@
+package me.danielschaefer.sensorwrangler.data
+
+enum class Picker {
+    FileOpen, FileSave, Directory, None
+}
