+++ src/me/danielschaefer/sensorwrangler/sensors/RandomSensor.kt
@@ -2,9 +2,9 @@ package me.danielschaefer.sensorwrangler.sensors
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import javafx.application.Platform
-import me.danielschaefer.sensorwrangler.Measurement
-import me.danielschaefer.sensorwrangler.NamedThreadFactory
 import me.danielschaefer.sensorwrangler.annotations.SensorProperty
+import me.danielschaefer.sensorwrangler.data.Measurement
+import me.danielschaefer.sensorwrangler.util.NamedThreadFactory
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledExecutorService
 import java.util.concurrent.ThreadLocalRandom
